pyproject.toml
[project]
name = "sqlllm-interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "google-genai>=1.49.0",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
    "sqlllm",
]
